+++ otx/algorithms/action/adapters/mmaction/task.py
@@ -158,8 +158,9 @@ class MMActionTask(OTXActionTask):
 
         recipe_cfg.work_dir = self._output_path
         recipe_cfg.resume = self._resume
-        recipe_cfg.distributed = False
         recipe_cfg.omnisource = False
+
+        self._configure_device(recipe_cfg, training)
 
         if data_cfg is not None:
             recipe_cfg.merge_from_dict(data_cfg)
@@ -195,6 +196,40 @@ class MMActionTask(OTXActionTask):
 
         self._config = recipe_cfg
         return recipe_cfg
+
+    def _configure_device(self, cfg: Config, training: bool):
+        """Setting device for training and inference."""
+        cfg.distributed = False
+        if torch.distributed.is_initialized():
+            cfg.gpu_ids = [int(os.environ["LOCAL_RANK"])]
+            if training:  # TODO multi GPU is available only in training. Evaluation needs to be supported later.
+                cfg.distributed = True
+                self.configure_distributed(cfg)
+        elif "gpu_ids" not in cfg:
+            gpu_ids = os.environ.get("CUDA_VISIBLE_DEVICES")
+            logger.info(f"CUDA_VISIBLE_DEVICES = {gpu_ids}")
+            if gpu_ids is not None:
+                cfg.gpu_ids = range(len(gpu_ids.split(",")))
+            else:
+                cfg.gpu_ids = range(1)
+
+        # consider "cuda" and "cpu" device only
+        if not torch.cuda.is_available():
+            cfg.device = "cpu"
+            cfg.gpu_ids = range(-1, 0)
+        else:
+            cfg.device = "cuda"
+
+    @staticmethod
+    def configure_distributed(cfg: Config):
+        """Patching for distributed training."""
+        if hasattr(cfg, "dist_params") and cfg.dist_params.get("linear_scale_lr", False):
+            new_lr = len(cfg.gpu_ids) * cfg.optimizer.lr
+            logger.info(
+                f"enabled linear scaling rule to the learning rate. \
+                changed LR from {cfg.optimizer.lr} to {new_lr}"
+            )
+            cfg.optimizer.lr = new_lr
 
     # pylint: disable=too-many-branches, too-many-statements
     def _train_model(