+++ otx/algorithms/segmentation/configs/ocr_lite_hrnet_18/model.py
@@ -22,10 +22,10 @@ _base_ = [
 ]
 
 model = dict(
-    type="ClassIncrEncoderDecoder",
+    type="OTXEncoderDecoder",
     pretrained=None,
     decode_head=dict(
-        type="FCNHead",
+        type="CustomFCNHead",
         in_channels=[40, 80, 160, 320],
         in_index=[0, 1, 2, 3],
         input_transform="multiple_select",
@@ -38,7 +38,6 @@ model = dict(
         norm_cfg=dict(type="BN", requires_grad=True),
         align_corners=False,
         enable_aggregator=True,
-        enable_out_norm=False,
         loss_decode=[
             dict(
                 type="CrossEntropyLoss",