+++ otx/algorithms/segmentation/configs/ocr_lite_hrnet_x_mod3/model.py
@@ -22,10 +22,10 @@ _base_ = [
 ]
 
 model = dict(
-    type="ClassIncrEncoderDecoder",
+    type="OTXEncoderDecoder",
     pretrained=None,
     decode_head=dict(
-        type="FCNHead",
+        type="CustomFCNHead",
         in_channels=[18, 60, 80, 160, 320],
         in_index=[0, 1, 2, 3, 4],
         input_transform="multiple_select",
@@ -41,8 +41,6 @@ model = dict(
         aggregator_min_channels=60,
         aggregator_merge_norm=None,
         aggregator_use_concat=False,
-        enable_out_norm=False,
-        enable_loss_equalizer=True,
         loss_decode=[
             dict(
                 type="CrossEntropyLoss",