+++ tests/integration/cli/classification/test_classification.py
@@ -122,7 +122,7 @@ class TestMultiClassClassificationCLI:
     @pytest.mark.parametrize("dump_features", [True, False])
     def test_otx_export(self, template, tmp_dir_path, dump_features):
         tmp_dir_path = tmp_dir_path / "multi_class_cls"
-        otx_export_testing(template, tmp_dir_path, dump_features)
+        otx_export_testing(template, tmp_dir_path, dump_features, check_ir_meta=True)
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)