+++ CHANGELOG.md
@@ -7,10 +7,11 @@ All notable changes to this project will be documented in this file.
 ### New features
 
 - Support direct annotation input for COCO format (<https://github.com/openvinotoolkit/training_extensions/pull/1921>)
+- Action task supports multi GPU training. (<https://github.com/openvinotoolkit/training_extensions/pull/2057>)
 
 ### Enhancements
 
--
+- Make semantic segmentation OpenVINO models compatible with ModelAPI (<https://github.com/openvinotoolkit/training_extensions/pull/2029>).
 
 ### Bug fixes